[refactor] move/room dto -> entity 수정
